--- a/core/types/log.go
+++ b/core/types/log.go
@@ -63,6 +63,9 @@ type logMarshaling struct {
 	Index       hexutil.Uint
 }
 
+//go:generate go run ../../rlp/rlpgen -type rlpLog -out gen_log_rlp.go
+
+// rlpLog is used to RLP-encode both the consensus and storage formats.
 type rlpLog struct {
 	Address common.Address
 	Topics  []common.Hash